Production info is its own entity seperate from anthology linked by id
