+++ apps/backend/src/seeds/production-info.seed.ts
@@ -1,0 +1,212 @@
+interface ProductionInfoSeedData {
+  anthologyTitle: string;
+  binding_type?: string;
+  printed_by?: string;
+  print_run?: number;
+  printing_cost?: number;
+  page_count?: number;
+  cover_image_file_link?: string;
+  design_files_link?: string;
+  dimensions?: string;
+  weight_in_grams?: number;
+}
+export const ProductionInfoSeed: ProductionInfoSeedData[] = [
+  {
+    anthologyTitle: 'Voices From the Threshold',
+    binding_type: 'Saddle Stitch',
+    printed_by: 'FlashPrint',
+    print_run: 60,
+    printing_cost: 198.0,
+    page_count: 48,
+    cover_image_file_link: 'publications/covers/voices-from-the-threshold.png',
+  },
+  {
+    anthologyTitle: 'The Color of Saturday',
+    binding_type: 'Tape Bound/Binding Strips',
+    printed_by: 'FlashPrint',
+    print_run: 80,
+    printing_cost: 224.4,
+    page_count: 36,
+    cover_image_file_link: 'publications/covers/color-of-saturday.png',
+  },
+  {
+    anthologyTitle: 'What the River Carries',
+    binding_type: 'Tape Bound/Binding Strips',
+    printed_by: 'FlashPrint',
+    print_run: 50,
+    printing_cost: 165.0,
+    page_count: 44,
+    cover_image_file_link: 'publications/covers/what-the-river-carries.png',
+  },
+  {
+    anthologyTitle: 'Prism Literary Magazine #14: Borrowed and Stolen',
+    binding_type: 'Perfect Bound',
+    printed_by: 'Country Press',
+    print_run: 75,
+    printing_cost: 681.75,
+    page_count: 88,
+    cover_image_file_link: 'publications/covers/rubix-14-front.png',
+  },
+  {
+    anthologyTitle: 'Snapshots at 3AM',
+    binding_type: 'Perfect Bound',
+    printed_by: 'PaperGraphics',
+    print_run: 320,
+    printing_cost: 1814.4,
+    page_count: 72,
+    cover_image_file_link: 'publications/covers/snapshots-3am.png',
+  },
+  {
+    anthologyTitle: 'How to Survive a Cafeteria',
+    binding_type: 'Tape Bound/Binding Strips',
+    printed_by: 'FlashPrint',
+    print_run: 100,
+    printing_cost: 180.0,
+    page_count: 40,
+    cover_image_file_link: 'publications/covers/how-to-survive-cafeteria.png',
+  },
+  {
+    anthologyTitle: 'Letters to Nobody',
+    binding_type: 'Tape Bound/Binding Strips',
+    printed_by: 'FlashPrint',
+    print_run: 80,
+    printing_cost: 249.6,
+    page_count: 52,
+    cover_image_file_link: 'publications/covers/letters-to-nobody.png',
+  },
+  {
+    anthologyTitle: 'Aftershock',
+    binding_type: 'Perfect Bound',
+    printed_by: 'PaperGraphics',
+    print_run: 500,
+    printing_cost: 2975.0,
+    page_count: 112,
+    cover_image_file_link: 'publications/covers/aftershock-cover.png',
+  },
+  {
+    anthologyTitle: 'Bright Noise',
+    binding_type: 'Perfect Bound',
+    printed_by: 'Country Press',
+    print_run: 100,
+    printing_cost: 856.0,
+    page_count: 96,
+    cover_image_file_link: 'publications/covers/bright-noise-cover.png',
+  },
+  {
+    anthologyTitle: "Tomorrow's Almanac",
+    binding_type: 'Perfect Bound',
+    printed_by: 'PaperGraphics',
+    print_run: 320,
+    printing_cost: 1830.4,
+    page_count: 76,
+    cover_image_file_link: 'publications/covers/tomorrows-almanac.png',
+  },
+  {
+    anthologyTitle: 'Prism Literary Magazine #15: Memory Palace',
+    binding_type: 'Perfect Bound',
+    printed_by: 'Country Press',
+    print_run: 75,
+    printing_cost: 697.5,
+    page_count: 92,
+    cover_image_file_link: 'publications/covers/rubix-15-front.png',
+  },
+  {
+    anthologyTitle: 'Not Guilty',
+    binding_type: 'Perfect Bound',
+    printed_by: 'PaperGraphics',
+    print_run: 500,
+    printing_cost: 2950.0,
+    page_count: 108,
+    cover_image_file_link: 'publications/covers/not-guilty-cover.png',
+  },
+  {
+    anthologyTitle: 'Unlocked',
+    binding_type: 'Perfect Bound',
+    printed_by: 'Marquis',
+    print_run: 500,
+    printing_cost: 2950.0,
+    page_count: 128,
+    cover_image_file_link: 'publications/covers/unlocked-cover.png',
+  },
+  {
+    anthologyTitle: 'Burn and Bloom',
+    binding_type: 'Perfect Bound',
+    printed_by: 'PaperGraphics',
+    print_run: 500,
+    printing_cost: 2900.0,
+    page_count: 116,
+    cover_image_file_link: 'publications/covers/burn-and-bloom-cover.png',
+  },
+  {
+    anthologyTitle: 'Civic Creatures',
+    binding_type: 'Perfect Bound',
+    printed_by: 'PaperGraphics',
+    print_run: 500,
+    printing_cost: 2975.0,
+    page_count: 120,
+    cover_image_file_link: 'publications/covers/civic-creatures-cover.png',
+  },
+  {
+    anthologyTitle: 'The Weight of a Suitcase',
+    binding_type: 'Saddle Stitch',
+    printed_by: 'FlashPrint',
+    print_run: 60,
+    printing_cost: 192.0,
+    page_count: 46,
+    cover_image_file_link: 'publications/covers/weight-of-suitcase.png',
+  },
+  {
+    anthologyTitle: 'Every Map Lies',
+    binding_type: 'Perfect Bound',
+    printed_by: 'Marquis',
+    print_run: 200,
+    printing_cost: 1380.0,
+    page_count: 78,
+    cover_image_file_link: 'publications/covers/every-map-lies.png',
+  },
+  {
+    anthologyTitle: 'The Night Kitchen',
+    binding_type: 'Tape Bound/Binding Strips',
+    printed_by: 'FlashPrint',
+    print_run: 50,
+    printing_cost: 180.0,
+    page_count: 46,
+    cover_image_file_link: 'publications/covers/night-kitchen.png',
+  },
+  {
+    anthologyTitle: 'Hard Pivot',
+    binding_type: 'Perfect Bound',
+    printed_by: 'Marquis',
+    print_run: 400,
+    printing_cost: 2400.0,
+    page_count: 122,
+    cover_image_file_link: 'publications/covers/hard-pivot.png',
+  },
+  {
+    anthologyTitle: 'The Space Between Languages',
+    binding_type: 'Tape Bound/Binding Strips',
+    printed_by: 'FlashPrint',
+    print_run: 50,
+    printing_cost: 187.5,
+    page_count: 48,
+    cover_image_file_link: 'publications/covers/space-between-languages.png',
+  },
+  {
+    anthologyTitle: 'Hallway Dispatches',
+    binding_type: 'Saddle Stitch',
+    printed_by: 'FlashPrint',
+    print_run: 40,
+    printing_cost: 48.0,
+    page_count: 24,
+    cover_image_file_link: 'publications/covers/hallway-dispatches.png',
+  },
+  {
+    anthologyTitle: 'Field Notes From the T',
+    binding_type: 'Saddle Stitch',
+    printed_by: 'FlashPrint',
+    print_run: 50,
+    printing_cost: 55.0,
+    page_count: 20,
+    cover_image_file_link: 'publications/covers/field-notes-from-the-t.png',
+  },
+];
