+++ apps/backend/src/seeds/seed-anthologies.ts
@@ -2,6 +2,7 @@ import { DataSource } from 'typeorm';
 import { Anthology } from 'src/anthology/anthology.entity';
 import { ProductionInfo } from 'src/production-info/production-info.entity';
 import { AnthologiesSeed } from './anthologies.seed';
+import { ProductionInfoSeed } from './production-info.seed';
 
 export async function seedAnthologies(dataSource: DataSource) {
   const anthologyRepo = dataSource.getRepository(Anthology);
@@ -9,24 +10,45 @@ export async function seedAnthologies(dataSource: DataSource) {
 
   console.log('Seeding anthologies...');
 
-  for (const { productionInfo: piData, ...data } of AnthologiesSeed) {
+  for (const data of AnthologiesSeed) {
     const exists = await anthologyRepo.findOne({
       where: { title: data.title },
     });
 
     if (!exists) {
-      const pi = piRepo.create(piData);
-      const savedPi = await piRepo.save(pi);
-
-      const entity = anthologyRepo.create({ ...data, productionInfo: savedPi });
-      const savedAnthology = await anthologyRepo.save(entity);
-
-      savedPi.anthology = savedAnthology;
-      await piRepo.save(savedPi);
-
+      const savedAnthology = await anthologyRepo.save(
+        anthologyRepo.create(data),
+      );
       console.log(`  ✓ Created anthology: ${savedAnthology.title}`);
     } else {
       console.log(`  - Anthology already exists: ${data.title}`);
+    }
+  }
+
+  console.log('Seeding production info...');
+
+  for (const { anthologyTitle, ...piData } of ProductionInfoSeed) {
+    const anthology = await anthologyRepo.findOne({
+      where: { title: anthologyTitle },
+    });
+
+    if (!anthology) {
+      console.warn(
+        `  ✗ Anthology not found for production info: ${anthologyTitle}`,
+      );
+      continue;
+    }
+
+    const exists = await piRepo.findOne({
+      where: { anthology: { id: anthology.id } },
+    });
+
+    if (!exists) {
+      const pi = piRepo.create({ ...piData, anthology });
+      await piRepo.save(pi);
+      console.log(`  ✓ Created production info for: ${anthologyTitle}`);
+    } else {
+      console.log(`  - Production info already exists for: ${anthologyTitle}`);
     }
   }
 }
