+++ apps/backend/src/seeds/story-drafts.seed.ts
@@ -2,6 +2,7 @@ import { EditRound, SubmissionRound } from '../story-draft/types';
 
 export interface StoryDraftSeedItem {
   storyTitle: string;
+  anthologyTitle: string;
   authorId: number;
   docLink: string;
   submissionRound: SubmissionRound;
@@ -10,13 +11,13 @@ export interface StoryDraftSeedItem {
   editRound: EditRound;
   proofread: boolean;
   notes: string[];
-  classPeriod: string;
 }
 
 export const StoryDraftsSeed: StoryDraftSeedItem[] = [
   {
     storyTitle: 'Standing at the Threshold',
     authorId: 1,
+    anthologyTitle: 'Voices From the Threshold',
     docLink: 'https://docs.google.com/document/d/1aStandingAtTheThreshold',
     submissionRound: SubmissionRound.ONE,
     studentConsent: true,
@@ -24,10 +25,10 @@ export const StoryDraftsSeed: StoryDraftSeedItem[] = [
     editRound: EditRound.TWO,
     proofread: true,
     notes: [],
-    classPeriod: 'Edwards 1/6',
   },
   {
     storyTitle: 'Saturday Mornings',
+    anthologyTitle: 'The Color of Saturday',
     authorId: 2,
     docLink: 'https://docs.google.com/document/d/2aSaturdayMornings',
     submissionRound: SubmissionRound.TWO,
@@ -36,10 +37,10 @@ export const StoryDraftsSeed: StoryDraftSeedItem[] = [
     editRound: EditRound.ONE,
     proofread: false,
     notes: ['Needs formatting review', 'Check dialogue punctuation'],
-    classPeriod: 'Edwards 1/6',
   },
   {
     storyTitle: 'The River Remembers',
+    anthologyTitle: '',
     authorId: 3,
     docLink: 'https://docs.google.com/document/d/3aTheRiverRemembers',
     submissionRound: SubmissionRound.ONE,
@@ -48,10 +49,10 @@ export const StoryDraftsSeed: StoryDraftSeedItem[] = [
     editRound: EditRound.TWO,
     proofread: true,
     notes: [],
-    classPeriod: 'Shin 3/5',
   },
   {
     storyTitle: 'Borrowed Words',
+    anthologyTitle: 'What the River Carries',
     authorId: 4,
     docLink: 'https://docs.google.com/document/d/4aBorrowedWords',
     submissionRound: SubmissionRound.THREE,
@@ -60,10 +61,10 @@ export const StoryDraftsSeed: StoryDraftSeedItem[] = [
     editRound: EditRound.ONE,
     proofread: false,
     notes: ['Awaiting parental consent form'],
-    classPeriod: 'Fuentes 2/4',
   },
   {
     storyTitle: '3:17 AM',
+    anthologyTitle: 'What the River Carries',
     authorId: 5,
     docLink: 'https://docs.google.com/document/d/5a317AM',
     submissionRound: SubmissionRound.TWO,
@@ -72,6 +73,5 @@ export const StoryDraftsSeed: StoryDraftSeedItem[] = [
     editRound: EditRound.TWO,
     proofread: true,
     notes: ['Final version approved by author'],
-    classPeriod: 'Nakamura 5/7',
   },
 ];
